+++ packages/create-rsbuild/package.json
@@ -29,8 +29,8 @@
     "registry": "https://registry.npmjs.org/"
   },
   "scripts": {
-    "build": "rslib --syntax es2023",
-    "dev": "rslib --syntax es2023 -w",
+    "build": "rs lib --syntax es2023",
+    "dev": "rs lib --syntax es2023 -w",
     "start": "node ./dist/index.js"
   },
   "dependencies": {
@@ -44,7 +44,6 @@
     "@rsbuild/plugin-solid": "workspace:*",
     "@rsbuild/plugin-svelte": "workspace:*",
     "@rsbuild/plugin-vue": "workspace:*",
-    "@rslib/core": "catalog:",
     "@types/node": "catalog:",
     "typescript": "catalog:"
   },
