--- packages/plugin-babel/rslib.config.ts
@@ -1,3 +1,0 @@
-import { pureEsmPackage } from '@scripts/config/rslib.config.ts';
-
-export default pureEsmPackage;
