+++ packages/plugin-less/package.json
@@ -29,8 +29,8 @@
     "registry": "https://registry.npmjs.org/"
   },
   "scripts": {
-    "build": "rslib",
-    "dev": "rslib -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "prebundle": "prebundle",
     "test": "rs test"
   },
@@ -42,7 +42,6 @@
   },
   "devDependencies": {
     "@rsbuild/core": "workspace:*",
-    "@rslib/core": "catalog:",
     "@scripts/test-helper": "workspace:*",
     "@types/less": "catalog:",
     "prebundle": "catalog:",
