--- packages/plugin-less/rslib.config.ts
@@ -1,9 +1,0 @@
-import { pureEsmPackage } from '@scripts/config/rslib.config.ts';
-import { defineConfig } from '@rslib/core';
-
-export default defineConfig({
-  ...pureEsmPackage,
-  output: {
-    externals: /[\\/]compiled[\\/]/,
-  },
-});
