+++ packages/plugin-sass/package.json
@@ -29,8 +29,8 @@
     "registry": "https://registry.npmjs.org/"
   },
   "scripts": {
-    "build": "rslib",
-    "dev": "rslib -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "prebundle": "prebundle",
     "test": "rs test"
   },
@@ -43,7 +43,6 @@
   },
   "devDependencies": {
     "@rsbuild/core": "workspace:*",
-    "@rslib/core": "catalog:",
     "@scripts/test-helper": "workspace:*",
     "@types/node": "catalog:",
     "@types/sass-loader": "catalog:",
