+++ packages/plugin-solid/package.json
@@ -29,8 +29,8 @@
     "registry": "https://registry.npmjs.org/"
   },
   "scripts": {
-    "build": "rslib",
-    "dev": "rslib -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "test": "rs test"
   },
   "dependencies": {
@@ -40,7 +40,6 @@
   },
   "devDependencies": {
     "@rsbuild/core": "workspace:*",
-    "@rslib/core": "catalog:",
     "@scripts/test-helper": "workspace:*",
     "@types/babel__core": "catalog:",
     "typescript": "catalog:"
