+++ packages/plugin-solid/rstack.config.ts
@@ -1,7 +1,24 @@
 import { define } from 'rstack';
 
+define.lib(async () => {
+  const { esmConfig, nodeMinifyConfig } = await import('@scripts/config/lib');
+
+  return {
+    lib: [
+      esmConfig,
+      {
+        format: 'cjs',
+        syntax: 'es2023',
+        output: {
+          minify: nodeMinifyConfig,
+        },
+      },
+    ],
+  };
+});
+
 define.test(async () => {
-  const { baseConfig } = await import('@scripts/config/rstest.config.ts');
+  const { baseConfig } = await import('@scripts/config/test');
 
   return baseConfig;
 });
