+++ packages/plugin-svelte/package.json
@@ -28,8 +28,8 @@
     "registry": "https://registry.npmjs.org/"
   },
   "scripts": {
-    "build": "rslib",
-    "dev": "rslib -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "test": "rs test"
   },
   "dependencies": {
@@ -38,7 +38,6 @@
   },
   "devDependencies": {
     "@rsbuild/core": "workspace:*",
-    "@rslib/core": "catalog:",
     "@scripts/test-helper": "workspace:*",
     "@types/node": "catalog:",
     "svelte": "catalog:",
