+++ packages/plugin-svgr/rstack.config.ts
@@ -1,7 +1,30 @@
 import { define } from 'rstack';
 
+define.lib(async () => {
+  const { esmConfig } = await import('@scripts/config/lib');
+
+  return {
+    lib: [
+      esmConfig,
+      {
+        source: {
+          entry: {
+            assetLoader: './src/assetLoader.ts',
+            loader: './src/loader.ts',
+          },
+        },
+        output: {
+          filename: {
+            js: '[name].mjs',
+          },
+        },
+      },
+    ],
+  };
+});
+
 define.test(async () => {
-  const { baseConfig } = await import('@scripts/config/rstest.config.ts');
+  const { baseConfig } = await import('@scripts/config/test');
 
   return baseConfig;
 });
