+++ pnpm-workspace.yaml
@@ -32,7 +32,6 @@ catalog:
   '@rsbuild/plugin-check-syntax': '^2.0.0'
   '@rsbuild/plugin-rem': '^1.0.6'
   '@rsbuild/plugin-type-check': '^1.5.0'
-  '@rslib/core': '1.0.0-beta.0'
   '@rspack/core': '~2.1.5'
   '@rspack/plugin-preact-refresh': '^2.0.1'
   '@rspack/plugin-react-refresh': '^2.0.2'
@@ -44,7 +43,6 @@ catalog:
   '@rstack-dev/doc-ui': '1.14.5'
   '@rstackjs/load-config': '^0.1.2'
   '@rstackjs/test-utils': '^0.2.0'
-  '@rstest/adapter-rslib': '^0.11.2'
   '@rstest/playwright': '^0.11.2'
   '@shikijs/transformers': '^4.3.1'
   '@svgr/core': '8.1.0'
